+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Added
+
+ - `CharSlice99_alloca_c_str`, which allocates a C-string using `alloca`.
+
 ## 0.7.3 - 2022-02-25
 
 ### Fixed